+++ tools/dataMigration/migrateData/migrateTablesData.ts
@@ -8,13 +8,15 @@ import { Objects } from '../../../src/utils/objects'
 import { DBNames } from '../_DBNames'
 import { getCreateViewDDL } from './_createDataViews'
 import { _getNodeInserts } from './_getNodeInserts'
-import { _getNodeInsertsDegradedForest } from './_getNodeInsertsDegradedForest'
+// import { _getNodeInsertsDegradedForest } from './_getNodeInsertsDegradedForest'
 import { isBasicTable } from './_repos'
-import { getNodeInsertsTableWithODP } from './getNodeInsertsTableWithODP'
+import { migrateFRATablesData } from './migrateFRATablesData'
+// import { getNodeInsertsTableWithODP } from './getNodeInsertsTableWithODP'
 
 export const migrateTablesData = async (
   props: { assessment: Assessment; cycle: Cycle },
-  client: ITask<any>
+  client: ITask<any>,
+  legacySchema = '_legacy'
 ): Promise<void> => {
   const { assessment, cycle } = props
   const schema = DBNames.getAssessmentSchema(assessment.props.name)
@@ -42,91 +44,20 @@ export const migrateTablesData = async (
       }
     }
   )
-  const tableDegradedForest = tables.find((t) => t.props.name === 'degradedForest')
-  const tableExtentOfForest = tables.find((t) => t.props.name === 'extentOfForest')
-  const tableForestCharacteristics = tables.find((t) => t.props.name === 'forestCharacteristics')
-  const tableGrowingStockAvg = tables.find((t) => t.props.name === 'growingStockAvg')
-  const tableGrowingStockTotal = tables.find((t) => t.props.name === 'growingStockTotal')
 
   // get node insert values
   const values = (
     await Promise.all(
       tables
         .filter((table) => isBasicTable(table.props.name))
-        .map(async (table) => _getNodeInserts({ assessment, cycle, countryISOs, table }, client))
+        .map(async (table) => _getNodeInserts({ assessment, cycle, countryISOs, table, legacySchema }, client))
     )
   ).flat()
 
-  // non basic tables insert
-  values.push(
-    ...(await _getNodeInsertsDegradedForest({ assessment, cycle, countryISOs, table: tableDegradedForest }, client))
-  )
-  values.push(
-    ...(await getNodeInsertsTableWithODP(
-      {
-        assessment,
-        table: tableExtentOfForest,
-        tableNameLegacy: 'eof_fra_values',
-        variables: ['forestArea', 'otherWoodedLand'],
-        estimated: true,
-      },
-      client
-    ))
-  )
-  values.push(
-    ...(await getNodeInsertsTableWithODP(
-      {
-        assessment,
-        table: tableForestCharacteristics,
-        tableNameLegacy: 'foc_fra_values',
-        variables: [
-          'naturalForestArea',
-          'plantationForestArea',
-          'plantationForestIntroducedArea',
-          'otherPlantedForestArea',
-        ],
-        estimated: true,
-      },
-      client
-    ))
-  )
+  if (!schema.includes('pan_european')) {
+    await migrateFRATablesData({ assessment, cycle, countryISOs, tables, values }, client)
+  }
 
-  values.push(
-    ...(await getNodeInsertsTableWithODP(
-      {
-        assessment,
-        table: tableGrowingStockAvg,
-        tableNameLegacy: 'growing_stock_avg',
-        variables: [
-          'naturallyRegeneratingForest',
-          'plantationForest',
-          'otherPlantedForest',
-          'otherWoodedLand',
-          'plantedForest',
-          'forest',
-        ],
-      },
-      client
-    ))
-  )
-  values.push(
-    ...(await getNodeInsertsTableWithODP(
-      {
-        assessment,
-        table: tableGrowingStockTotal,
-        tableNameLegacy: 'growing_stock_total',
-        variables: [
-          'naturallyRegeneratingForest',
-          'plantationForest',
-          'otherPlantedForest',
-          'otherWoodedLand',
-          'plantedForest',
-          'forest',
-        ],
-      },
-      client
-    ))
-  )
   // insert nodes
   const schemaCycle = DBNames.getCycleSchema(assessment.props.name, cycle.name)
   const pgp = pgPromise()
@@ -135,7 +66,6 @@ export const migrateTablesData = async (
   })
   const query = pgp.helpers.insert(values, cs)
   await client.none(query)
-
   // create data views
   const queries = await Promise.all(tables.map((table) => getCreateViewDDL({ assessment, cycle, table })))
